+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/DocDisplay.java
@@ -339,4 +339,5 @@ public interface DocDisplay extends HasValueChangeHandlers<Void>,
    boolean onInsertSnippet();
    
    Position getDocumentEnd();
+   void setSurroundSelectionPref(String value);
 }