+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/ace/AceEditorNative.java
@@ -507,5 +507,9 @@ public class AceEditorNative extends JavaScriptObject {
       uiPrefsSynced_ = true;
    }
    
+   public final native void setSurroundSelectionPref(String value) /*-{
+      this.$surroundSelection = value;
+   }-*/;
+   
    private static boolean uiPrefsSynced_ = false;
 }